+++ app/controllers/claims_controller.rb
@@ -6,6 +6,7 @@ class ClaimsController < ApplicationController
 
   def create
     claim = Claim.create!(eligibility: StudentLoans::Eligibility.new)
+
     session[:claim_id] = claim.to_param
 
     redirect_to claim_path("qts-year")